backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-ai-inference==1.0.0b9
azure-common==1.1.28
azure-core==1.37.0
azure-cosmos==4.14.3
azure-identity==1.25.1
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
jiter==0.12.0
Levenshtein==0.27.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
oauthlib==3.3.1
openai==2.14.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.21
RapidFuzz==3.14.3
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
starlette==0.50.0
thefuzz==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
yarl==1.22.0
